+++ annabaas-java/src/org/domcode/talk/raffler/annaffler/util/IProcessor.java
@@ -1,0 +1,5 @@
+package org.domcode.talk.raffler.annaffler.util;
+
+public interface IProcessor {
+	public String process();
+}
