+++ annabaas-java/src/org/domcode/talk/raffler/annaffler/util/ITask.java
@@ -1,0 +1,7 @@
+package org.domcode.talk.raffler.annaffler.util;
+
+public interface ITask {
+
+	public void execute();
+	
+}
