+++ annabaas-java/src/org/domcode/talk/raffler/annaffler/util/IWriter.java
@@ -1,0 +1,8 @@
+package org.domcode.talk.raffler.annaffler.util;
+
+public interface IWriter {
+	public void write(String stringToWrite);
+	public void hookFunctionOne();
+	public void hookFunctionTwo();
+	public void hookFunctionThree();
+}
