--- src/org/domcode/talk/raffler/annaffler/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author abaas
- *
- */
-package org.domcode.talk.raffler.annaffler;
