--- src/org/domcode/talk/raffler/annaffler/util/IParticipantData.java
@@ -1,10 +1,0 @@
-package org.domcode.talk.raffler.annaffler.util;
-
-public interface IParticipantData {
-	
-	public void setName(String name);
-	public String getName();
-	public void setParticipantId(Integer participantId);
-	public Integer getParticipantId();
-
-}
