--- src/org/domcode/talk/raffler/annaffler/util/IProcessor.java
@@ -1,5 +1,0 @@
-package org.domcode.talk.raffler.annaffler.util;
-
-public interface IProcessor {
-	public String process();
-}
